+++ .pipelines/nuget.yml
@@ -1,8 +1,10 @@
 parameters:
 - name: RunOrtExtensionsTests
+  # this is only nuget packing pipeline, we don't want to run test again for that
+  # we should have run the test somewhere already.
   displayName: Run Tests?
   type: boolean
-  default: true
+  default: false
 
 - name: DoCompliance
   displayName: Run Compliance Tasks?
@@ -30,13 +32,10 @@ pr:
       - "*"
       
 stages:
-- template: templates/c-api-cpu.yml
+- template: templates/build-package-for-nuget.yml
   parameters:
     RunOrtExtensionsTests: ${{ parameters.RunOrtExtensionsTests }}
     DoCompliance: ${{ parameters.DoCompliance }}
     DoEsrp: ${{ parameters.DoEsrp }}
     IsReleaseBuild: ${{ parameters.IsReleaseBuild }}
     OrtNugetPackageId: 'Microsoft.ML.OnnxRuntime.Extensions'
-    AdditionalBuildFlags: ''
-    AdditionalWinBuildFlags: ''
-    BuildVariant: 'default'
