+++ .pipelines/templates/build-package-for-macosx.yml
@@ -1,0 +1,59 @@
+parameters:
+- name: AdditionalBuildFlags
+  displayName: Additional build flags for build.sh
+  type: string
+  default: ''
+- name: IsReleaseBuild
+  displayName: "Is this is a release build?"
+  type: boolean
+  default: false
+
+- name: WithCache
+  displayName: Build with Cache
+  type: boolean
+  default: false
+
+stages:
+- stage: MacOS_C_API_Packaging_CPU
+  dependsOn: []
+  jobs:
+  - template: mac-cpu-packing-jobs.yml
+    parameters:
+      MacosArch: 'x86_64'
+      AdditionalBuildFlags: ${{ parameters.AdditionalBuildFlags }} 
+      WithCache: ${{ parameters.WithCache }}
+  - template: mac-cpu-packing-jobs.yml
+    parameters:
+      MacosArch: 'arm64'
+      AdditionalBuildFlags: "${{ parameters.AdditionalBuildFlags }} -DCMAKE_OSX_ARCHITECTURES=arm64"
+      WithCache: ${{ parameters.WithCache }}
+  - template: mac-cpu-packing-jobs.yml
+    parameters:
+      MacosArch: 'universal2'
+      AdditionalBuildFlags: "${{ parameters.AdditionalBuildFlags }} -DCMAKE_OSX_ARCHITECTURES=\"arm64;x86_64\""
+      WithCache: ${{ parameters.WithCache }}
+  - job: MacOS_C_API_Package_Publish_All
+    dependsOn:
+    - MacOS_C_API_Packaging_CPU_x86_64
+    - MacOS_C_API_Packaging_CPU_arm64
+    - MacOS_C_API_Packaging_CPU_universal2
+    pool:
+      vmImage: 'macOS-12'
+    steps:
+    - task: DownloadPipelineArtifact@2
+      inputs:
+        artifact: 'onnxruntime-extensions-osx-x86_64'
+        targetPath: '$(Build.ArtifactStagingDirectory)'
+    - task: DownloadPipelineArtifact@2
+      inputs:
+        artifact: 'onnxruntime-extensions-osx-arm64'
+        targetPath: '$(Build.ArtifactStagingDirectory)'
+    - task: DownloadPipelineArtifact@2
+      inputs:
+        artifact: 'onnxruntime-extensions-osx-universal2'
+        targetPath: '$(Build.ArtifactStagingDirectory)'
+    - task: PublishPipelineArtifact@1
+      inputs:
+        targetPath: '$(Build.ArtifactStagingDirectory)'
+        artifact: 'onnxruntime-extensions-osx'
+        condition: 'succeededOrFailed()'
