+++ .pipelines/templates/mac-cpu-packing-jobs.yml
@@ -17,14 +17,15 @@ parameters:
   type: string
   default: ''
 
-# Must be 1 or 0
-- name: AllowReleasedOpsetOnly
-  displayName: Whether unreleased onnx opsets are allowed
-  type: number
-  default: 1
-  values:
-  - 1
-  - 0
+- name: CacheDir
+  displayName: Cache Directory
+  type: string
+  default: ''
+
+- name: Today
+  type: string
+  default: ""
+
 
 jobs:
 - job: MacOS_C_API_Packaging_CPU_${{ parameters.MacosArch }}
@@ -32,7 +33,6 @@ jobs:
     clean: all
   variables:
     MACOSX_DEPLOYMENT_TARGET: '10.14'
-    ALLOW_RELEASED_ONNX_OPSET_ONLY: ${{ parameters.AllowReleasedOpsetOnly }}
     TODAY: $[format('{0:dd}{0:MM}{0:yyyy}', pipeline.startTime)]
   pool:
     vmImage: 'macOS-12'
@@ -46,34 +46,50 @@ jobs:
     inputs:
       versionSpec: '16.x'
 
-  #- template: set-version-number-variables-step.yml
+  - template: set-package-version-variable-step.yml
+    parameters:
+      IsReleaseBuild: true
+      PackageVersionVariableName: ORT_EXTENSIONS_VERSION
 
-  - ${{ if eq(parameters.MacosArch, 'universal2') }}:
-    - template: mac-cpu-packaging-steps.yml
-      parameters:
-        MacosArch: ${{ parameters.MacosArch }}
-        AdditionalBuildFlags: ${{ parameters.AdditionalBuildFlags }} --cmake_extra_defines CMAKE_OSX_ARCHITECTURES="arm64;x86_64"
-        WithCache: ${{ parameters.WithCache }}
-        ${{ if eq(parameters.WithCache, true) }}:
-          Today: $(TODAY)
-          CacheDir: $(ORT_CACHE_DIR)
+  - template: mac-build-step-with-cache.yml
+    parameters:
+      WithCache: ${{ parameters.WithCache }}
+      Today: ${{ parameters.Today }}
+      AdditionalKey: onnxruntime-extensions_${{ parameters.MacosArch }}
+      CacheDir: ${{ parameters.CacheDir }}
+      ChangeEveryCommit: true
+      BuildStep:
+        - script: |
+            set -e -x
+            cd $(Build.SourcesDirectory)
+            rm -rf out/
+            ./build.sh -DOCOS_ENABLE_CV2=OFF -DOCOS_ENABLE_BLINGFIRE=OFF ${{parameters.AdditionalBuildFlags}}
+            ls out/*
+          displayName: 'Build ${{ parameters.MacosArch }}'
+          env:
+            CCACHE_DIR: ${{ parameters.CacheDir }}
 
-  - ${{ if eq(parameters.MacosArch, 'arm64') }}:
-    - template: mac-cpu-packaging-steps.yml
-      parameters:
-        MacosArch: ${{ parameters.MacosArch }}
-        AdditionalBuildFlags: ${{ parameters.AdditionalBuildFlags }} --cmake_extra_defines CMAKE_OSX_ARCHITECTURES=arm64
-        WithCache: ${{ parameters.WithCache }}
-        ${{ if eq(parameters.WithCache, true) }}:
-          Today: $(TODAY)
-          CacheDir: $(ORT_CACHE_DIR)
+  - task: Bash@3
+    displayName: 'Copy build artifacts for zipping'
+    inputs:
+      targetType: 'inline'
+      script: |
+        set -e -x
+        target_lib_path=onnxruntime-extensions-osx-${{ parameters.MacosArch }}-$(ORT_EXTENSIONS_VERSION)/lib
+        mkdir -p $target_lib_path
+        cp $(Build.SourcesDirectory)/out/Darwin/RelWithDebInfo/lib/libortextensions.dylib  ${target_lib_path}/
+        strip -S ${target_lib_path}/libortextensions.dylib
+      workingDirectory: '$(Build.BinariesDirectory)'
 
-  - ${{ if eq(parameters.MacosArch, 'x86_64') }}:
-    - template: mac-cpu-packaging-steps.yml
-      parameters:
-        MacosArch: ${{ parameters.MacosArch }}
-        AdditionalBuildFlags: ${{ parameters.AdditionalBuildFlags }}
-        WithCache: ${{ parameters.WithCache }}
-        ${{ if eq(parameters.WithCache, true) }}:
-          Today: $(TODAY)
-          CacheDir: $(ORT_CACHE_DIR)
+  - task: ArchiveFiles@2
+    inputs:
+      rootFolderOrFile: '$(Build.BinariesDirectory)/onnxruntime-extensions-osx-${{ parameters.MacosArch }}-$(ORT_EXTENSIONS_VERSION)'
+      includeRootFolder: true
+      archiveType: 'tar' # Options: zip, 7z, tar, wim
+      tarCompression: 'gz'
+      archiveFile: '$(Build.ArtifactStagingDirectory)/onnxruntime-extensions-osx-${{ parameters.MacosArch }}.tgz'
+      replaceExistingArchive: true
+
+
+  - publish: '$(Build.ArtifactStagingDirectory)'
+    artifact: 'onnxruntime-extensions-osx-${{ parameters.MacosArch }}'
