+++ nuget/NativeNuget.nuspec
@@ -33,6 +33,24 @@
 		<file src="LICENSE.txt" target="LICENSE.txt" />
 		<!-- Windows, Linux, macOS via netstandard -->
 		<file src="..\out\Windows\bin\RelWithDebInfo\ortextensions.dll" target="runtimes\win-x64\native" />
+		
+		<!-- It's safe to add more packages here, even it's not existed in CI pipeline. Those files would be removed autumaticlly -->
+		<file src="..\includes\*.h" target="build\native\include" />
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-android\onnxruntime-extensions.aar" target="runtimes/android/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-linux-aarch64\lib\libortextensions.so" target="runtimes/linux-arm64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-linux-x64\lib\libortextensions.so" target="runtimes/linux-x64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-osx-arm64\lib\libortextensions.dylib" target="runtimes/osx.10.14-arm64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-osx-x86_64\lib\libortextensions.dylib" target="runtimes/osx.10.14-x64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-arm\lib\ortextensions.dll" target="runtimes/win-arm/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-arm\lib\ortextensions.lib" target="runtimes/win-arm/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-arm64\lib\ortextensions.dll" target="runtimes/win-arm64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-arm64\lib\ortextensions.lib" target="runtimes/win-arm64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-x64\lib\ortextensions.dll" target="runtimes/win-x64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-x64\lib\ortextensions.lib" target="runtimes/win-x64/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-x86\lib\ortextensions.dll" target="runtimes/win-x86/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime-extensions-win-x86\lib\ortextensions.lib" target="runtimes/win-x86/native"/>
+		<file src="..\nuget-artifacts-ext\onnxruntime_extensions.xcframework\**" target="runtimes/ios/native"/>
+
 		<file src="targets\netstandard\Microsoft.ML.OnnxRuntime.Extensions.props" target="build\native" />
 		<file src="targets\netstandard\Microsoft.ML.OnnxRuntime.Extensions.props" target="build\netstandard1.1" />
 		<file src="targets\netstandard\Microsoft.ML.OnnxRuntime.Extensions.props" target="build\netstandard2.0" />
